backend/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai>=0.3.0
